Fixing second type mismatch

+++ stra.c
@@ -54,10 +54,9 @@ in either array. */
 int Str_compare(const char s1[], const char s2[]) {
     size_t i = 0;
     assert(s1 != NULL && s2 != NULL);
-
     while (s1[i] != '\0' && s2[i] != '\0') {
         if (s1[i] != s2[i]) {
-            return s1[i] - s2[i];
+            return (int)(s1[i] - s2[i]);
         }
         i++;
     }
